resize-service/requirements.txt
pika==1.3.2
boto3==1.35.0
Pillow==12.0.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-mock==3.14.0
